refactor: Position 관련 클래스 position 패키지로 이동
