+++ src/main/java/janggi/domain/BoardState.java
@@ -1,5 +1,7 @@
 package janggi.domain;
 
+import janggi.domain.position.Position;
+
 public interface BoardState {
     boolean hasPieceAt(Position position);
     Piece getPieceAt(Position position);
