+++ src/main/java/janggi/domain/movestrategy/ChaMoveStrategy.java
@@ -1,9 +1,9 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.BoardState;
-import janggi.domain.Column;
-import janggi.domain.Position;
-import janggi.domain.Row;
+import janggi.domain.position.Column;
+import janggi.domain.position.Position;
+import janggi.domain.position.Row;
 
 public class ChaMoveStrategy implements MoveStrategy {
