+++ src/main/java/janggi/domain/movestrategy/GungseongBoundMoveStrategy.java
@@ -1,7 +1,7 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.BoardState;
-import janggi.domain.Position;
+import janggi.domain.position.Position;
 import janggi.domain.Team;
 
 public class GungseongBoundMoveStrategy implements MoveStrategy {
