+++ src/main/java/janggi/domain/movestrategy/MaMoveStrategy.java
@@ -1,9 +1,9 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.BoardState;
-import janggi.domain.Column;
-import janggi.domain.Position;
-import janggi.domain.Row;
+import janggi.domain.position.Column;
+import janggi.domain.position.Position;
+import janggi.domain.position.Row;
 
 import java.util.ArrayList;
 import java.util.List;
