+++ src/main/java/janggi/domain/movestrategy/MoveStrategy.java
@@ -1,7 +1,7 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.BoardState;
-import janggi.domain.Position;
+import janggi.domain.position.Position;
 
 public interface MoveStrategy {
     boolean canMove(Position from, Position to, BoardState boardState);
