+++ src/main/java/janggi/domain/movestrategy/PoMoveStrategy.java
@@ -1,11 +1,11 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.BoardState;
-import janggi.domain.Column;
+import janggi.domain.position.Column;
 import janggi.domain.Piece;
 import janggi.domain.PieceType;
-import janggi.domain.Position;
-import janggi.domain.Row;
+import janggi.domain.position.Position;
+import janggi.domain.position.Row;
 
 public class PoMoveStrategy implements MoveStrategy {
