+++ src/main/java/janggi/domain/position/Position.java
@@ -1,4 +1,4 @@
-package janggi.domain;
+package janggi.domain.position;
 
 import java.util.HashMap;
 import java.util.Map;
