+++ src/test/java/janggi/domain/PieceTest.java
@@ -2,6 +2,9 @@ package janggi.domain;
 
 import janggi.domain.movestrategy.JolMoveStrategy;
 import janggi.domain.movestrategy.MoveStrategy;
+import janggi.domain.position.Column;
+import janggi.domain.position.Position;
+import janggi.domain.position.Row;
 import janggi.exception.InvalidMoveException;
 import org.junit.jupiter.api.Test;
